Update the span_id format (and parent_span_id)
V2 of Google Cloud Stackdriver api uses an 16-character hexadecimal encoding of an 8-byte array for the Span Id (contrary to v1, where it is a 64 bit integer). (see https://cloud.google.com/trace/docs/reference/v2/rest/v2/projects.traces/batchWrite#Span) 
But the implementation of the stackdriver exporter was still using an integer as span ID.
This change consists of converting the integer span Id to an hexadecimal string.

This is also now consistent with the implementation of the HTTP Trace Context Propagator.