Fix jaeger_exporter to properly handle root spans and timestamps
Currently exporting to jaeger doesn't work due to two problems:
1. root span has a null parent id, therefore thirft message sent
by jaeger_exporter.Collector to jaeger results in silent failure to
deserialize as parent span id is unset but is a required field.

2. It constructed start/end timestamp with second precision expanded to
milisecond units. Spec requires them to be expanded to microseconds

This PR fixes:
1. by setting parent_span_id to 0 (per jaeger docs) if missing.

2. by expanding timestamp to microsecond precision by multiplying it by
1e6 instead of 1e3 and adding dt.microseconds. This mirrors existing
implementation in zipkin exporter

Fixes: #185