Add TraceLoggingFilter.
Filters are another, flexible way to add contextual information to the logs.

A filter can be attached to both a Logger and a Handler, allowing for more granular control on how to add the fields. E.g. a filter can be attached only to a specific opencensus export handler, it can be added anytime in the runtime. And this will affect all loggers using this handler, whether they were constructed before or after adding a filter (unlike with a custom Logger or Adapter).